Add files via upload
ii: From the three images, I can tell that the maps are very similar, which means most of the complaints are fairly accurate; however, there were lots of fireworks on Staten Island but the complaints were relatively low. Maybe the people on Staten Island have a higher tolerance for fireworks.

iii: Having data collected from a longer period, not just a month, can increase the accuracy of the data, which makes the data visualization more truthful. Or use colors that can identify the differences clearly, for instance, using red and green to increase the difference visually.